requirements.txt
aiohttp>=3.9.0
openai>=1.12.0
pandas>=2.1.0
numpy>=1.26.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
APScheduler>=3.10.0
